refactor(js): split adapter dependency seams
